requirements.txt
google-cloud-pubsub>=2.18.0
python-dotenv>=1.0.0

